Game of Life
For this assignment I am hoping to build an android app using processing through the Eclipse IDE.


The game will be to survive as long as possible based on decisions you make every year. 
The game will calculate your chance of death based on mortality rates for each age you make it to. 
The game will have different resource bars based on a range of lifestyle parameters such as health, wealth and happiness.
These parameters can improve your chances of surviving a certain cause of death such as for example, heart disease can be reduced by having a healthy character.
I am hoping to publish it on the android appstore as it is something i would like to have on my phone to play.