Changed functionality of game to drastically alter purpose
I decided that I didnt like how my original idea was turning out. My
original idea was to have questions on life decisions alter how a
characters stats turned out. I could only think of a few questions that
would alter the characters stats.

While i was in the process of creating this game i chose to make a
database of questions that i wanted to be as customizable as possible.
So as i was unhappy with the mechanics of how my original game worked i
decided to change this game into a trivia game. With the database i
created it is easy to add any type of questions and answers and the
program will still function.

So after a bit of renaming of variables and removal of the randomizers i
have changed this project from a specific game about decisions to a
program that can handle any number of questions filled into a database.
This program is now a tool that can handle fun things such as trivia
quiz games or something terrible like a multiple choice test